Added location into the ranking algorithm for the home feed.

+++ app/src/main/java/com/codepath/travelbud/Post.java
@@ -47,7 +47,7 @@ public class Post extends ParseObject {
         put(KEY_USER, user);
     }
 
-    public Integer getRating() { return getInt(KEY_RATING); }
+    public float getRating() { return (float) getDouble(KEY_RATING); }
 
     public void setRating(Float rating) { put(KEY_RATING, rating); }
